requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Differential privacy
opacus>=1.0.0
torch-dp>=0.2.0

# Testing and development
pytest>=7.0.0
pytest-cov>=3.0.0
black>=22.0.0
isort>=5.10.0
flake8>=4.0.0
mypy>=0.950

# Documentation
sphinx>=4.0.0
sphinx-rtd-theme>=1.0.0

# Utilities
tqdm>=4.62.0
pyyaml>=6.0
tensorboard>=2.8.0
wandb>=0.12.0  # Optional: for experiment tracking
